🌱 FEATURE | fix mistakes and clean names

+++ SDK/SDK/Models/Directives/float.cs
@@ -6,9 +6,9 @@ using System.Threading.Tasks;
 
 namespace SDK.Models.Directives
 {
-    public class float : BaseModel
+    public class float_ : BaseModel
     {
-        public float() : base(".float")
+        public float_() : base(".float")
         {
             this.Comment = "The .float directive generates a single-precision floating-point constant into the current section. The .float directive is not valid in the .bss section.";
         }
